// src/app/common/models/Customer.ts
export class Customer {
    id: string;
    first_name: string;
    last_name: string;
    organization_id: string;
    organization_name: string;
    organization_number: string;
    address: string;
    apt_num: string;
    city: string;
    state: string;
    zip: string;
    phone: string;
    additional_phone: string;
    email: string;
    additional_email: string;
    owner_id: string;
    create_date: Date;
    update_date: Date;
}